ゲームマネージャーに時間とUIとコイン枚数追加（仮） シーンロード追加（未呼び出し） 各ソース整頓 シーンタイトルコントローラ対応

+++ source/BulletManager.h
@@ -14,7 +14,7 @@ private:
 private:
 	//	定数
 	static const int BULLET_MAX = 201;		//	弾最大数
-	const int LIMIT = 60 * 1.5f;
+	const int LIMIT = ( const int )( 60 * 1.5f );
 
 	//	オブジェクト
 	iexMesh*	org;			//	オリジナルモデル
